requirements.txt
flask==3.0.3
python-dotenv==1.0.1
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
pydantic<2